% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arctos_data.R
\name{get_arctos_data}
\alias{get_arctos_data}
\title{get_arctos_data}
\usage{
get_arctos_data(scientific_name, api_key)
}
\arguments{
\item{scientific_name}{Scientific or Latin name for a species}

\item{api_key}{Your API key used to access Arctos}
}
\value{
dataframe
}
\description{
Returns a dataframe of
}
