% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_code_tables.R
\name{list_code_tables}
\alias{list_code_tables}
\title{list_code_tables}
\usage{
list_code_tables(api_key)
}
\arguments{
\item{api_key}{Your API key used to access Arctos}
}
\value{
list
}
\description{
Returns a list of code tables used by Arctos
}
