% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_code_table.R
\name{request_code_table}
\alias{request_code_table}
\title{request_code_table}
\usage{
request_code_table(table, api_key)
}
\arguments{
\item{table}{Name of the Arctos table to query}

\item{api_key}{Your API key used to access Arctos}
}
\value{
list
}
\description{
Returns more detail about a code table in Arctos
}
